requirements.txt
google-cloud-bigquery>=3.0
google-cloud-storage>=2.0
requests>=2.31
python-dotenv>=1.0
